fix(grafana): Complete SSO provisioning and org setup
PROBLEM:
- Iframe showed login page (302 redirect) instead of dashboard
- Dashboard returned 404 (user in wrong org)
- Missing provisioning logs and diagnostic tools

SOLUTION:

1. Backend (AuthRequestController):
   - Enhanced ensureUser/ensureOrgMembership/ensureActiveOrg with Step 1/3, 2/3, 3/3 logs
   - Fixed API calls: lookupUser() returns Optional<UserLookupResponse>
   - Added START/SUCCESS provisioning logs with emojis
   - Added imports: CreateUserResponse, UserLookupResponse, Optional

2. Frontend (grafanaUrl.ts):
   - Removed url.searchParams.set('orgId', String(orgId))
   - Backend setUserActiveOrg handles org selection (no URL param needed)
   - orgId parameter kept for API compatibility but unused

3. Grafana (grafana.ini):
   - Fixed jwk_set_url: http://backend:8080/.well-known/jwks.json (internal Docker network)
   - Verified JWT config: enabled=true, header_name=X-Org-JWT, org_id_claim=orgId, auto_sign_up=true
   - Verified security: allow_embedding=true, disable_login_form=true

4. Nginx (nginx-ssl.conf.template):
   - Verified /_auth/grafana has proxy_set_header Cookie $http_cookie (line 92)
   - Verified /core-admin/monitoring/ has proxy_set_header X-Org-JWT (line 119)
   - No changes needed - config is correct

5. Diagnostic CLI:
   - Created scripts/grafana/user-check.sh <login> (lookup user, list orgs, show current org)
   - Created scripts/grafana/user-set-org.sh <login> <orgId> (manually set active org)
   - Both use Grafana Admin API with basic auth

6. CI Integration:
   - Created scripts/grafana/sanity-test.sh (tests test_admin org 2 membership, current org, dashboard 200)
   - Added Makefile target: make test-grafana-sso
   - Added to help-advanced menu

RESULT:
✅ Iframe loads dashboard without login prompt
✅ Logs show 3 provisioning steps (Step 1/3, 2/3, 3/3)
✅ user-check.sh test_admin shows org 2 membership and current org 2
✅ CI sanity test passes: make test-grafana-sso

KEY CHANGE:
Backend uses POST /api/users/{userId}/using/{orgId} (plural 'users' - admin endpoint)
NOT /api/user/using/{orgId} (singular - would affect current user, not target user)

FILES CHANGED:
- backend/src/main/java/cz/muriel/core/monitoring/AuthRequestController.java
- frontend/src/utils/grafanaUrl.ts
- docker/grafana/grafana.ini
- scripts/grafana/user-check.sh (new)
- scripts/grafana/user-set-org.sh (new)
- scripts/grafana/sanity-test.sh (new)
- Makefile
- GRAFANA_SSO_FIX_SUMMARY.md (new)

TESTING:
# Local
make dev-up
make test-grafana-sso
open https://admin.core-platform.local/monitoring

# CI
make up && make test-grafana-sso

DOCUMENTATION:
See GRAFANA_SSO_FIX_SUMMARY.md for full architecture, debugging guide, and examples.

+++ backend/src/main/java/cz/muriel/core/monitoring/GrafanaJwtService.java
@@ -142,7 +142,7 @@ public class GrafanaJwtService {
       }
 
       // Map tenant to Grafana org
-      int grafanaOrgId = tenantRegistry.getGrafanaOrgId(tenantId);
+      long grafanaOrgId = tenantRegistry.getGrafanaOrgId(tenantId);
 
       // Map Keycloak roles to Grafana role
       String grafanaRole = mapKeycloakRoleToGrafanaFromJwt(kcToken);
@@ -156,10 +156,10 @@ public class GrafanaJwtService {
       JWSHeader header = new JWSHeader.Builder(JWSAlgorithm.RS256).keyID(keyProvider.getKeyId())
           .build();
 
-      // Build JWT claims
+      // Build JWT claims - orgId MUST be numeric (Long) for Grafana org_id_claim
       JWTClaimsSet claims = new JWTClaimsSet.Builder().issuer(issuer).subject(username)
           .claim("email", email).claim("name", name != null ? name : username)
-          .claim("preferred_username", username).claim("orgId", grafanaOrgId)
+          .claim("preferred_username", username).claim("orgId", Long.valueOf(grafanaOrgId))
           .claim("role", grafanaRole).issueTime(Date.from(now)).expirationTime(Date.from(expiry))
           .jwtID(jti).build();
